backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.11.0
